feat(nextjs): improve support for custom server
- Add new generator for custom server
- Add @nrwl/js:node executor for serving custom server
- Deprecate @nrwl/node:node in favor of the above
