+++ packages/js/src/executors/node/compat.ts
@@ -1,0 +1,4 @@
+import { convertNxExecutor } from '@nrwl/devkit';
+import nodeExecutor from './node.impl';
+
+export default convertNxExecutor(nodeExecutor);
