+++ packages/next/src/generators/custom-server/schema.d.ts
@@ -1,0 +1,4 @@
+export interface CustomServerSchema {
+  project: string;
+  compiler?: 'tsc' | 'swc';
+}
