+++ packages/next/src/generators/custom-server/schema.json
@@ -1,0 +1,34 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "cli": "nx",
+  "$id": "NxNextCustomServer",
+  "title": "Add custom server",
+  "description": "Add a custom server to existing Next.js application.",
+  "examples": [
+    {
+      "command": "nx g custom-server my-app",
+      "description": "Add a custom server to existing Next.js app."
+    }
+  ],
+  "type": "object",
+  "properties": {
+    "project": {
+      "type": "string",
+      "description": "The name of the project.",
+      "alias": "p",
+      "$default": {
+        "$source": "argv",
+        "index": 0
+      },
+      "x-dropdown": "project",
+      "x-prompt": "What is the name of the project to set up a custom server for?"
+    },
+    "compiler": {
+      "type": "string",
+      "enum": ["tsc", "swc"],
+      "default": "tsc",
+      "description": "The compiler used to build the custom server."
+    }
+  },
+  "required": ["project"]
+}
